+++ src/ui/dialogs/importdata.cpp
@@ -212,7 +212,7 @@ void ImportData::LoadFile()
             model->setEditable(true);
             m_table->setModel(model);
             if (model->columnCount() == 2 && model->rowCount() > 100)
-                QMessageBox::warning(this, QString("Whow!"), QString("This rather long xy file should probably be treated as thermogram. Just push the Import Thermogram on left.\nBut please be aware that, the automatic peak picking will probably fail to import the data correctly."));
+                QMessageBox::warning(this, QString("Whow!"), QString("This rather long xy file should probably be treated as thermogram. Just push the Import Thermogram on left.\nBut please be aware that, the automatic peak picking will probably fail to import the data correctly.\nYou need the time between each inject and the starting time for the first injection."));
             NoChanged();
         } else {
             QMessageBox::warning(this, QString("File not supported!"), QString("Sorry, but I don't know this format. Try a simple table."));