+++ Changelog.md
@@ -1,5 +1,6 @@
 # SupraFit 2.0 pre-Alpha
 
+- Import of thermogram from plain txt files possible, peak time and peak start have to be given for peak picking
 - Improved thermodynamic output -> calculation of free enthalpy and entropy (if heat available) including confidence
 - ToolSet::Confidence functions gives identical results compared to the octave quantile function
 - More consistent parameter names